semaine1/requirements.txt
gradio==3.40.0
Pillow==9.4.0
opencv-python-headless==4.8.0.74
matplotlib==3.7.1
numpy==1.24.3
